chore: move splitPrimitiveProps to utils folder
