+++ packages/react/src/primitives/Expander/Expander.tsx
@@ -7,7 +7,7 @@ import { useDeprecationWarning } from '../../hooks/useDeprecationWarning';
 import { ComponentClassNames } from '../shared/constants';
 import { ExpanderProps } from '../types/expander';
 import { Primitive } from '../types/view';
-import { splitPrimitiveProps } from '../shared/styleUtils';
+import { splitPrimitiveProps } from '../utils/splitPrimitiveProps';
 
 const ExpanderPrimitive: Primitive<ExpanderProps, typeof Root> = (
   {
