+++ packages/react/src/primitives/shared/styleUtils.ts
@@ -1,8 +1,6 @@
 import * as React from 'react';
 
 import {
-  AllStyleProps,
-  BaseStyleProps,
   ComponentPropsToStylePropsMap,
   GridItemStyleProps,
   GridSpanType,
@@ -16,7 +14,6 @@ import { Breakpoint, Breakpoints } from '../types/responsive';
 
 import { useTheme } from '../../hooks';
 import { isEmptyString, isNullOrEmptyString } from './utils';
-import { FlexContainerStyleProps } from '../types/flex';
 import { ThemeStylePropKey } from '../types/theme';
 import { WebTheme } from '@aws-amplify/ui';
 
@@ -163,50 +160,4 @@ export const useStyles = (
       }),
     [propStyles, style, breakpoints, breakpoint, tokens]
   );
-};
-
-interface SplitProps<PrimitiveProps> {
-  styleProps: AllStyleProps;
-  rest: Partial<
-    Omit<PrimitiveProps, keyof FlexContainerStyleProps | keyof BaseStyleProps>
-  >;
-}
-
-function isStyleKey(prop: string): prop is keyof AllStyleProps {
-  return prop in ComponentPropsToStylePropsMap;
-}
-
-/**
- * This function splits props into style props and non-style props. This is used
- * on Field primitives so we can apply style props on the wrapper element and
- * the rest on the input.
- * @param props this should be a destructured `rest` from the component's props
- */
-export const splitPrimitiveProps = <
-  PrimitiveProps extends Record<string, unknown>
->(
-  props: PrimitiveProps
-): SplitProps<PrimitiveProps> => {
-  const splitProps: SplitProps<PrimitiveProps> = {
-    styleProps: {},
-    rest: {},
-  };
-
-  Object.keys(props).forEach((prop) => {
-    if (isStyleKey(prop)) {
-      // we know it is a style key
-      // so we know we can assign the key in styleProps
-      splitProps.styleProps = {
-        ...splitProps.styleProps,
-        [prop]: props[prop],
-      };
-    } else {
-      splitProps.rest = {
-        ...splitProps.rest,
-        [prop]: props[prop],
-      };
-    }
-  });
-
-  return splitProps;
 };
